requirements.txt
streamlit
python-dotenv
python-docx
python-pptx
pdfplumber
google-generativeai
